Refactor: Avatar 컴포넌트를 common으로 이동
