+++ src/components/common/common.styles.ts
@@ -1,0 +1,12 @@
+export const styles = {
+  // Avatar styles
+  avatar: 'rounded-full flex items-center justify-center',
+
+  // Layout styles
+  flexCenter: 'flex items-center justify-center',
+  flexBetween: 'flex items-center justify-between',
+
+  // Text styles
+  textSemibold: 'font-semibold',
+  textSmall: 'text-sm',
+}
