+++ src/components/common/index.ts
@@ -1,1 +1,3 @@
 export { default as Header } from './Header'
+export { default as Avatar } from './Avatar'
+export { styles } from './common.styles'
